+++ core/lib.deno_core.d.ts
@@ -184,5 +184,13 @@ declare namespace Deno {
       after_hook?: (promise: Promise<unknown>) => void,
       resolve_hook?: (promise: Promise<unknown>) => void,
     ): void;
+
+    const build: {
+      target: string;
+      arch: string;
+      os: string;
+      vendor: string;
+      env: string | undefined;
+    };
   }
 }
