+++ runtime/js/06_util.js
@@ -1,18 +1,10 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-const core = globalThis.Deno.core;
-const internals = globalThis.__bootstrap.internals;
 const primordials = globalThis.__bootstrap.primordials;
 const {
-  decodeURIComponent,
-  ObjectPrototypeIsPrototypeOf,
   Promise,
   SafeArrayIterator,
-  SafeRegExp,
-  StringPrototypeReplace,
-  TypeError,
 } = primordials;
-import { URLPrototype } from "internal:deno_url/00_url.js";
 let logDebug = false;
 let logSource = "JS";
 
@@ -45,64 +37,6 @@ function createResolvable() {
   promise.reject = reject;
   return promise;
 }
-
-// Keep in sync with `fromFileUrl()` in `std/path/win32.ts`.
-function pathFromURLWin32(url) {
-  let p = StringPrototypeReplace(
-    url.pathname,
-    new SafeRegExp(/^\/*([A-Za-z]:)(\/|$)/),
-    "$1/",
-  );
-  p = StringPrototypeReplace(
-    p,
-    /\//g,
-    "\\",
-  );
-  p = StringPrototypeReplace(
-    p,
-    new SafeRegExp(/%(?![0-9A-Fa-f]{2})/g),
-    "%25",
-  );
-  let path = decodeURIComponent(p);
-  if (url.hostname != "") {
-    // Note: The `URL` implementation guarantees that the drive letter and
-    // hostname are mutually exclusive. Otherwise it would not have been valid
-    // to append the hostname and path like this.
-    path = `\\\\${url.hostname}${path}`;
-  }
-  return path;
-}
-
-// Keep in sync with `fromFileUrl()` in `std/path/posix.ts`.
-function pathFromURLPosix(url) {
-  if (url.hostname !== "") {
-    throw new TypeError(`Host must be empty.`);
-  }
-
-  return decodeURIComponent(
-    StringPrototypeReplace(
-      url.pathname,
-      new SafeRegExp(/%(?![0-9A-Fa-f]{2})/g),
-      "%25",
-    ),
-  );
-}
-
-function pathFromURL(pathOrUrl) {
-  if (ObjectPrototypeIsPrototypeOf(URLPrototype, pathOrUrl)) {
-    if (pathOrUrl.protocol != "file:") {
-      throw new TypeError("Must be a file URL.");
-    }
-
-    return core.build.os == "windows"
-      ? pathFromURLWin32(pathOrUrl)
-      : pathFromURLPosix(pathOrUrl);
-  }
-  return pathOrUrl;
-}
-
-// TODO(bartlomieju): remove
-internals.pathFromURL = pathFromURL;
 
 function writable(value) {
   return {
@@ -145,7 +79,6 @@ export {
   getterOnly,
   log,
   nonEnumerable,
-  pathFromURL,
   readOnly,
   setLogDebug,
   writable,
