+++ runtime/js/10_permissions.js
@@ -2,8 +2,8 @@
 
 const core = globalThis.Deno.core;
 const ops = core.ops;
+import { pathFromURL } from "internal:deno_web/00_infra.js";
 import { Event, EventTarget } from "internal:deno_web/02_event.js";
-import { pathFromURL } from "internal:runtime/06_util.js";
 const primordials = globalThis.__bootstrap.primordials;
 const {
   ArrayIsArray,
